backend/requirements.txt
annotated-types==0.7.0
anyio==4.3.0
asgiref==3.8.1
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.0
colorama==0.4.6
cryptography==44.0.2
distlib==0.3.8
distro==1.9.0
dj-rest-auth==7.0.1
Django==5.1.4
django-allauth==65.7.0
django-autoslug==1.9.9
django-cities-light==3.10.1
django-cleanup==8.1.0
django-cors-headers==4.6.0
django-filter==24.3
django-jazzmin==3.0.1
django-restframework==0.0.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
filelock==3.13.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.10
jwt==1.3.1
openai==1.30.3
pillow==11.0.0
platformdirs==4.2.0
progressbar2==4.5.0
psycopg2==2.9.9
pycparser==2.22
pydantic==2.7.1
pydantic_core==2.18.2
PyJWT==2.10.1
python-utils==3.9.1
pytz==2024.2
requests==2.32.3
setuptools==69.5.1
shurjopay-plugin==0.1.9
sniffio==1.3.1
sqlparse==0.5.2
tqdm==4.66.4
typing_extensions==4.12.2
tzdata==2024.2
Unidecode==1.3.8
urllib3==2.2.3
virtualenv==20.25.1
whitenoise==6.8.2
